+++ CHANGES.rst
@@ -5,7 +5,9 @@ grafana-wtf changelog
 
 in progress
 ===========
-- Prepare test suite for testing two different dashboard schema versions, v27 and v33
+- CI: Use most recent Grafana 8.3.3
+- Add two more examples about using `explore dashboards` with `jq`
+- CI: Prepare test suite for testing two different dashboard schema versions, v27 and v33
 - Improve determinism by returning stable sort order of dashboard results
 
 2021-12-11 0.12.0